handle not found project with a good error message

+++ switch.sh
@@ -20,9 +20,14 @@ function switch()
       source ~/.switch/`cat ~/.switch/proj.save`/out.sh
       rm ~/.switch/proj.save
     fi
-    echo "Switching to project \"$1\""
-    source ~/.switch/$1/in.sh
-    echo "$1" > ~/.switch/proj.save
+    if [[ -f ~/.switch/$1  ]]
+    then
+      echo "Switching to project \"$1\""
+      source ~/.switch/$1/in.sh
+      echo "$1" > ~/.switch/proj.save
+    else
+      echo "\"$1\" is not a switch project. Available projects: `ls -x ~/.switch`">&2
+    fi
   ;;
   * )
     echo "Too many arguments">&2
